+++ LiveHTS.Presentation/Interfaces/ViewModel/IClientContactViewModel.cs
@@ -10,7 +10,7 @@ namespace LiveHTS.Presentation.Interfaces.ViewModel
         IndexClientDTO IndexClientDTO { get; set; }
         ClientContactAddressDTO ContactAddress { get; set; }
         string ClientInfo { get; set; }
-        int? Telephone { get; set; }
+        long? Telephone { get; set; }
         string Landmark { get; set; }
         string PersonId { get; set; }
         string ContactId { get; set; }
